// frontend/src/shared/api/model/adminCompanyResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Tarikma API
 * Backend API for Tarik.ma
 * OpenAPI spec version: v1
 */

export interface AdminCompanyResponse {
  id?: number;
  name?: string;
  description?: string;
  verified?: boolean;
  createdAt?: string;
  memberCount?: number;
}